backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pandas==2.2.2
numpy==1.26.4
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
transformers==4.40.*
torch==2.3.*
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.*
slowapi==0.1.9
sqlalchemy==2.0.*
aiosqlite==0.20.*
httpx==0.27.*
pytest==8.3.*
pytest-asyncio==0.25.*
